+++ src/josepy/__init__.py
@@ -26,20 +26,14 @@ Originally developed as part of the ACME_ protocol implementation.
 
 """
 # flake8: noqa
-from josepy.b64 import (
-    b64decode,
-    b64encode,
-)
-
+from josepy.b64 import b64decode, b64encode
 from josepy.errors import (
     DeserializationError,
-    SerializationError,
     Error,
+    SerializationError,
     UnrecognizedTypeError,
 )
-
 from josepy.interfaces import JSONDeSerializable
-
 from josepy.json_util import (
     Field,
     JSONObjectWithFields,
@@ -54,37 +48,27 @@ from josepy.json_util import (
     encode_hex16,
     field,
 )
-
 from josepy.jwa import (
+    ES256,
+    ES384,
+    ES512,
     HS256,
     HS384,
     HS512,
-    JWASignature,
     PS256,
     PS384,
     PS512,
     RS256,
     RS384,
     RS512,
-    ES256,
-    ES384,
-    ES512,
+    JWASignature,
 )
-
-from josepy.jwk import (
-    JWK,
-    JWKRSA,
-)
-
-from josepy.jws import (
-    Header,
-    JWS,
-    Signature,
-)
-
+from josepy.jwk import JWK, JWKEC, JWKRSA, JWKOct
+from josepy.jws import JWS, Header, Signature
 from josepy.util import (
-    ComparableX509,
+    ComparableECKey,
     ComparableKey,
     ComparableRSAKey,
+    ComparableX509,
     ImmutableMap,
 )