+++ src/josepy/util.py
@@ -1,13 +1,13 @@
 """JOSE utilities."""
 import abc
-from types import ModuleType
-from typing import Union, Any, Callable, Iterator, Tuple, List, TypeVar, cast
-from collections.abc import Hashable, Mapping
 import sys
 import warnings
+from collections.abc import Hashable, Mapping
+from types import ModuleType
+from typing import Any, Callable, Iterator, List, Tuple, TypeVar, Union, cast
 
-from OpenSSL import crypto
 from cryptography.hazmat.primitives.asymmetric import ec, rsa
+from OpenSSL import crypto
 
 
 # Deprecated. Please use built-in decorators @classmethod and abc.abstractmethod together instead.