+++ tests/jwa_test.py
@@ -2,8 +2,9 @@
 import unittest
 from unittest import mock
 
-from josepy import errors
 import test_util
+
+from josepy import errors
 
 RSA256_KEY = test_util.load_rsa_private_key('rsa256_key.pem')
 RSA512_KEY = test_util.load_rsa_private_key('rsa512_key.pem')
@@ -50,8 +51,7 @@ class JWASignatureTest(unittest.TestCase):
         self.assertEqual(self.Sig2.to_partial_json(), 'Sig2')
 
     def test_from_json(self):
-        from josepy.jwa import JWASignature
-        from josepy.jwa import RS256
+        from josepy.jwa import RS256, JWASignature
         self.assertIs(JWASignature.from_json('RS256'), RS256)
 
 
@@ -75,8 +75,7 @@ class JWARSTest(unittest.TestCase):
         self.assertRaises(errors.Error, RS256.sign, RSA512_KEY.public_key(), b'foo')
 
     def test_sign_key_too_small(self):
-        from josepy.jwa import RS256
-        from josepy.jwa import PS256
+        from josepy.jwa import PS256, RS256
         self.assertRaises(errors.Error, RS256.sign, RSA256_KEY, b'foo')
         self.assertRaises(errors.Error, PS256.sign, RSA256_KEY, b'foo')
 
@@ -106,31 +105,11 @@ class JWARSTest(unittest.TestCase):
         RS256.sign(key, "message")
         self.assertIs(key.sign.called, True)
 
-    def test_sign_old_api(self):
-        from josepy.jwa import RS256
-        key = mock.MagicMock(spec=[u'signer'])
-        signer = mock.MagicMock()
-        key.signer.return_value = signer
-        RS256.sign(key, "message")
-        self.assertIs(key.signer.called, True)
-        self.assertIs(signer.update.called, True)
-        self.assertIs(signer.finalize.called, True)
-
     def test_verify_new_api(self):
         from josepy.jwa import RS256
         key = mock.MagicMock()
         RS256.verify(key, "message", "signature")
         self.assertIs(key.verify.called, True)
-
-    def test_verify_old_api(self):
-        from josepy.jwa import RS256
-        key = mock.MagicMock(spec=[u'verifier'])
-        verifier = mock.MagicMock()
-        key.verifier.return_value = verifier
-        RS256.verify(key, "message", "signature")
-        self.assertIs(key.verifier.called, True)
-        self.assertIs(verifier.update.called, True)
-        self.assertIs(verifier.verify.called, True)
 
 
 class JWAECTest(unittest.TestCase):
@@ -159,9 +138,10 @@ class JWAECTest(unittest.TestCase):
         self.assertIs(ES384.verify(EC_P384_KEY.public_key(), message, signature), False)
 
     def test_verify_with_different_key(self):
-        from josepy.jwa import ES256
-        from cryptography.hazmat.primitives.asymmetric import ec
         from cryptography.hazmat.backends import default_backend
+        from cryptography.hazmat.primitives.asymmetric import ec
+
+        from josepy.jwa import ES256
 
         message = b'foo'
         signature = ES256.sign(EC_P256_KEY, message)
@@ -169,55 +149,36 @@ class JWAECTest(unittest.TestCase):
         self.assertIs(ES256.verify(different_key.public_key(), message, signature), False)
 
     def test_sign_new_api(self):
-        from josepy.jwa import ES256
         from cryptography.hazmat.primitives.asymmetric.ec import SECP256R1
+
+        from josepy.jwa import ES256
         key = mock.MagicMock(curve=SECP256R1())
         with mock.patch("josepy.jwa.decode_dss_signature") as decode_patch:
             decode_patch.return_value = (0, 0)
             ES256.sign(key, "message")
         self.assertIs(key.sign.called, True)
 
-    def test_sign_old_api(self):
-        from josepy.jwa import ES256
-        from cryptography.hazmat.primitives.asymmetric.ec import SECP256R1
-        key = mock.MagicMock(spec=[u'signer'], curve=SECP256R1())
-        signer = mock.MagicMock()
-        key.signer.return_value = signer
-        with mock.patch("josepy.jwa.decode_dss_signature") as decode_patch:
-            decode_patch.return_value = (0, 0)
-            ES256.sign(key, "message")
-        self.assertIs(key.signer.called, True)
-        self.assertIs(signer.update.called, True)
-        self.assertIs(signer.finalize.called, True)
-
     def test_verify_new_api(self):
         import math
-        from josepy.jwa import ES256
+
         from cryptography.hazmat.primitives.asymmetric.ec import SECP256R1
+
+        from josepy.jwa import ES256
         key = mock.MagicMock(key_size=256, curve=SECP256R1())
         ES256.verify(key, "message", b'\x00' * math.ceil(key.key_size / 8) * 2)
         self.assertIs(key.verify.called, True)
-
-    def test_verify_old_api(self):
-        import math
-        from josepy.jwa import ES256
-        from cryptography.hazmat.primitives.asymmetric.ec import SECP521R1
-        key = mock.MagicMock(spec=[u'verifier'], key_size=521, curve=SECP521R1())
-        verifier = mock.MagicMock()
-        key.verifier.return_value = verifier
-        ES256.verify(key, "message", b'\x00' * math.ceil(key.key_size / 8) * 2)
-        self.assertIs(key.verifier.called, True)
-        self.assertIs(verifier.update.called, True)
-        self.assertIs(verifier.verify.called, True)
 
     def test_signature_size(self):
         from josepy.jwa import ES512
         from josepy.jwk import JWK
         key = JWK.from_json(
             {
-                'd': 'Af9KP6DqLRbtit6NS_LRIaCP_-NdC5l5R2ugbILdfpv6dS9R4wUPNxiGw-vVWumA56Yo1oBnEm8ZdR4W-u1lPHw5',
-                'x': 'AD4i4STyJ07iZJkHkpKEOuICpn6IHknzwAlrf-1w1a5dqOsRe30EECSN4vFxaeAmtdBSCKBwCq7h1q4bPgMrMUvF',
-                'y': 'AHAlXxrabjcx_yBxGObnm_DkEQMJK1E69OHY3x3VxF5VXoKc93CG4GLoaPvphZQvZnt5EfExQoPktwOMIVhBHaFR',
+                'd': 'Af9KP6DqLRbtit6NS_LRIaCP_-NdC5l5R2ugbILdfpv6dS9R4wUPNxiGw'
+                     '-vVWumA56Yo1oBnEm8ZdR4W-u1lPHw5',
+                'x': 'AD4i4STyJ07iZJkHkpKEOuICpn6IHknzwAlrf-1w1a5dqOsRe30EECSN4vFxae'
+                     'AmtdBSCKBwCq7h1q4bPgMrMUvF',
+                'y': 'AHAlXxrabjcx_yBxGObnm_DkEQMJK1E69OHY3x3VxF5VXoKc93CG4GLoaPvphZQv'
+                     'Znt5EfExQoPktwOMIVhBHaFR',
                 'crv': 'P-521',
                 'kty': 'EC'
             })