+++ tests/util_test.py
@@ -216,6 +216,7 @@ class frozendictTest(unittest.TestCase):  # pylint: disable=invalid-name
 
     def test_init_other_raises_type_error(self):
         from josepy.util import frozendict
+
         # specifically fail for generators...
         self.assertRaises(TypeError, frozendict, {'a': 'b'}.items())